+++ externs/wavefile.js
@@ -441,8 +441,9 @@ WaveFile.prototype.toSampleRate = function(
     sincFilterSize: 32,
     lanczosFilterSize: 24,
     sincWindow: function(x){},
-    LPForder: 1,
-    LPF: true}) {};
+    LPF: true,
+    LPFType: 'IIR',
+    LPForder: 1}) {};
 
 /**
  * Encode a 16-bit wave file as 4-bit IMA ADPCM.