+++ test/resampler-full/point.js
@@ -23,7 +23,7 @@ describe('Downsample a 32-bit 44.1kHz log sine sweep', function() {
     fs.readFileSync(path + "32fp-44100Hz-mono-chirp-1-22050-log.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(16000, {method: 'point', LPF: true});
+  wav.toSampleRate(16000, {method: 'point', LPF: true, LPFType: 'FIR'});
 
   // Write the file
   fs.writeFileSync(
@@ -37,7 +37,7 @@ describe('Upsample a 32-bit 44.1kHz log sine sweep', function() {
     fs.readFileSync(path + "32fp-44100Hz-mono-chirp-1-22050-log.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(96000, {method: 'point', LPF: true});
+  wav.toSampleRate(96000, {method: 'point', LPF: true, LPFType: 'FIR'});
 
   // Write the file
   fs.writeFileSync(
@@ -51,7 +51,7 @@ describe('Downsample a 32-bit 44.1kHz linear sine sweep', function() {
     fs.readFileSync(path + "32fp-44100Hz-mono-chirp-1-22050-linear.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(16000, {method: 'point', LPF: true});
+  wav.toSampleRate(16000, {method: 'point', LPF: true, LPFType: 'FIR'});
 
   // Write the file
   fs.writeFileSync(
@@ -65,7 +65,7 @@ describe('Upsample a 32-bit 44.1kHz linear sine sweep', function() {
     fs.readFileSync(path + "32fp-44100Hz-mono-chirp-1-22050-linear.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(96000, {method: 'point', LPF: true});
+  wav.toSampleRate(96000, {method: 'point', LPF: true, LPFType: 'FIR'});
 
   // Write the file
   fs.writeFileSync(
@@ -81,7 +81,7 @@ describe('Downsample a 16-bit 44.1kHz log sine sweep', function() {
     fs.readFileSync(path + "16bit-44100Hz-mono-chirp-1-22050-log.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(16000, {method: 'point', LPF: true});
+  wav.toSampleRate(16000, {method: 'point', LPF: true, LPFType: 'FIR'});
 
   // Write the file
   fs.writeFileSync(
@@ -95,7 +95,7 @@ describe('Upsample a 16-bit 44.1kHz log sine sweep', function() {
     fs.readFileSync(path + "16bit-44100Hz-mono-chirp-1-22050-log.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(96000, {method: 'point', LPF: true});
+  wav.toSampleRate(96000, {method: 'point', LPF: true, LPFType: 'FIR'});
 
   // Write the file
   fs.writeFileSync(
@@ -109,7 +109,7 @@ describe('Downsample a 16-bit 44.1kHz linear sine sweep', function() {
     fs.readFileSync(path + "16bit-44100Hz-mono-chirp-1-22050-linear.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(16000, {method: 'point', LPF: true});
+  wav.toSampleRate(16000, {method: 'point', LPF: true, LPFType: 'FIR'});
 
   // Write the file
   fs.writeFileSync(
@@ -123,7 +123,7 @@ describe('Upsample a 16-bit 44.1kHz linear sine sweep', function() {
     fs.readFileSync(path + "16bit-44100Hz-mono-chirp-1-22050-linear.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(96000, {method: 'point', LPF: true});
+  wav.toSampleRate(96000, {method: 'point', LPF: true, LPFType: 'FIR'});
 
   // Write the file
   fs.writeFileSync(
@@ -139,7 +139,7 @@ describe('Downsample a 16bit 44.1kHz file', function() {
     fs.readFileSync(path + "song1.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(16000, {method: 'point', LPF: true});
+  wav.toSampleRate(16000, {method: 'point', LPF: true, LPFType: 'FIR'});
 
   // Write the file
   fs.writeFileSync(
@@ -153,7 +153,7 @@ describe('Upsample a 16bit 44.1kHz file', function() {
     fs.readFileSync(path + "song1.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(96000, {method: 'point', LPF: true});
+  wav.toSampleRate(96000, {method: 'point', LPF: true, LPFType: 'FIR'});
 
   // Write the file
   fs.writeFileSync(