+++ test/resampler-full/sinc-IIR.js
@@ -2,7 +2,7 @@
 * WaveFile: https://github.com/rochars/wavefile
 * Copyright (c) 2017-2018 Rafael da Silva Rocha. MIT License.
 *
-* Tests for sample rate conversion, cubic method.
+* Tests for sample rate conversion, sinc method.
 * 
 */
 
@@ -11,7 +11,7 @@ const fs = require("fs");
 const WaveFile = require("../../test/loader.js");
 const path = "./test/files/";
 
-console.log('cubic');
+console.log('sinc + IIR');
 
 let hrstart = process.hrtime();
 
@@ -23,11 +23,11 @@ describe('Downsample a 32-bit 44.1kHz log sine sweep', function() {
     fs.readFileSync(path + "32fp-44100Hz-mono-chirp-1-22050-log.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(16000, {method: 'cubic', LPFType: 'FIR'});
+  wav.toSampleRate(16000, {method: 'sinc', LPFType: 'IIR'});
 
   // Write the file
   fs.writeFileSync(
-    path + "/out/to-sample-rate/cubic-32fp-16000Hz-mono-chirp-1-22050-log.wav",
+    path + "/out/to-sample-rate/sinc-IIR-32fp-16000Hz-mono-chirp-1-22050-log.wav",
     wav.toBuffer());
 });
 
@@ -37,11 +37,11 @@ describe('Upsample a 32-bit 44.1kHz log sine sweep', function() {
     fs.readFileSync(path + "32fp-44100Hz-mono-chirp-1-22050-log.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(96000, {method: 'cubic', LPFType: 'FIR'});
+  wav.toSampleRate(96000, {method: 'sinc', LPFType: 'IIR'});
 
   // Write the file
   fs.writeFileSync(
-    path + "/out/to-sample-rate/cubic-32fp-96000Hz-mono-chirp-1-22050-log.wav",
+    path + "/out/to-sample-rate/sinc-IIR-32fp-96000Hz-mono-chirp-1-22050-log.wav",
     wav.toBuffer());
 });
 
@@ -51,11 +51,11 @@ describe('Downsample a 32-bit 44.1kHz linear sine sweep', function() {
     fs.readFileSync(path + "32fp-44100Hz-mono-chirp-1-22050-linear.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(16000, {method: 'cubic', LPFType: 'FIR'});
+  wav.toSampleRate(16000, {method: 'sinc', LPFType: 'IIR'});
 
   // Write the file
   fs.writeFileSync(
-    path + "/out/to-sample-rate/cubic-32fp-16000Hz-mono-chirp-1-22050-linear.wav",
+    path + "/out/to-sample-rate/sinc-IIR-32fp-16000Hz-mono-chirp-1-22050-linear.wav",
     wav.toBuffer());
 });
 
@@ -65,11 +65,11 @@ describe('Upsample a 32-bit 44.1kHz linear sine sweep', function() {
     fs.readFileSync(path + "32fp-44100Hz-mono-chirp-1-22050-linear.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(96000, {method: 'cubic', LPFType: 'FIR'});
+  wav.toSampleRate(96000, {method: 'sinc', LPFType: 'IIR'});
 
   // Write the file
   fs.writeFileSync(
-    path + "/out/to-sample-rate/cubic-32fp-96000Hz-mono-chirp-1-22050-linear.wav",
+    path + "/out/to-sample-rate/sinc-IIR-32fp-96000Hz-mono-chirp-1-22050-linear.wav",
     wav.toBuffer());
 });
 
@@ -81,11 +81,11 @@ describe('Downsample a 16-bit 44.1kHz log sine sweep', function() {
     fs.readFileSync(path + "16bit-44100Hz-mono-chirp-1-22050-log.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(16000, {method: 'cubic', LPFType: 'FIR'});
+  wav.toSampleRate(16000, {method: 'sinc', LPFType: 'IIR'});
 
   // Write the file
   fs.writeFileSync(
-    path + "/out/to-sample-rate/cubic-16bit-16000Hz-mono-chirp-1-22050-log.wav",
+    path + "/out/to-sample-rate/sinc-IIR-16bit-16000Hz-mono-chirp-1-22050-log.wav",
     wav.toBuffer());
 });
 
@@ -95,11 +95,11 @@ describe('Upsample a 16-bit 44.1kHz log sine sweep', function() {
     fs.readFileSync(path + "16bit-44100Hz-mono-chirp-1-22050-log.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(96000, {method: 'cubic', LPFType: 'FIR'});
+  wav.toSampleRate(96000, {method: 'sinc', LPFType: 'IIR'});
 
   // Write the file
   fs.writeFileSync(
-    path + "/out/to-sample-rate/cubic-16bit-96000Hz-mono-chirp-1-22050-log.wav",
+    path + "/out/to-sample-rate/sinc-IIR-16bit-96000Hz-mono-chirp-1-22050-log.wav",
     wav.toBuffer());
 });
 
@@ -109,11 +109,11 @@ describe('Downsample a 16-bit 44.1kHz linear sine sweep', function() {
     fs.readFileSync(path + "16bit-44100Hz-mono-chirp-1-22050-linear.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(16000, {method: 'cubic', LPFType: 'FIR'});
+  wav.toSampleRate(16000, {method: 'sinc', LPFType: 'IIR'});
 
   // Write the file
   fs.writeFileSync(
-    path + "/out/to-sample-rate/cubic-16bit-16000Hz-mono-chirp-1-22050-linear.wav",
+    path + "/out/to-sample-rate/sinc-IIR-16bit-16000Hz-mono-chirp-1-22050-linear.wav",
     wav.toBuffer());
 });
 
@@ -123,11 +123,11 @@ describe('Upsample a 16-bit 44.1kHz linear sine sweep', function() {
     fs.readFileSync(path + "16bit-44100Hz-mono-chirp-1-22050-linear.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(96000, {method: 'cubic', LPFType: 'FIR'});
+  wav.toSampleRate(96000, {method: 'sinc', LPFType: 'IIR'});
 
   // Write the file
   fs.writeFileSync(
-    path + "/out/to-sample-rate/cubic-16bit-96000Hz-mono-chirp-1-22050-linear.wav",
+    path + "/out/to-sample-rate/sinc-IIR-16bit-96000Hz-mono-chirp-1-22050-linear.wav",
     wav.toBuffer());
 });
 
@@ -139,11 +139,11 @@ describe('Downsample a 16bit 44.1kHz file', function() {
     fs.readFileSync(path + "song1.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(16000, {method: 'cubic', LPFType: 'FIR'});
+  wav.toSampleRate(16000, {method: 'sinc', LPFType: 'IIR'});
 
   // Write the file
   fs.writeFileSync(
-    path + "/out/to-sample-rate/cubic-song1-16kHz.wav",
+    path + "/out/to-sample-rate/sinc-IIR-song1-16kHz.wav",
     wav.toBuffer());
 });
 
@@ -153,11 +153,11 @@ describe('Upsample a 16bit 44.1kHz file', function() {
     fs.readFileSync(path + "song1.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(96000, {method: 'cubic', LPFType: 'FIR'});
+  wav.toSampleRate(96000, {method: 'sinc', LPFType: 'IIR'});
 
   // Write the file
   fs.writeFileSync(
-    path + "/out/to-sample-rate/cubic-song1-96kHz.wav",
+    path + "/out/to-sample-rate/sinc-IIR-song1-96kHz.wav",
     wav.toBuffer());
 });